部分页面提交

+++ mioa-org/src/main/java/com/mjkj/mioa/org/dao/OrgPositionRepository.java
@@ -28,9 +28,9 @@ import com.mjkj.mioa.org.entity.TOrgPosition;
 public interface OrgPositionRepository extends JpaRepository<TOrgPosition, String>
 {
 
-	TOrgPosition findPositionByNameAndDomainAndIsdeleteEquals(String name,String domain, int i);
+	TOrgPosition findPositionByNameAndDomainAndIsdeleteEquals(String name,String domain, byte i);
 
-	List<TOrgPosition> findPositionByDomainAndIsdeleteEquals(String domain, int i);
+	List<TOrgPosition> findPositionByDomainAndIsdeleteEquals(String domain, byte i);
 	
 }
